update:     cors 이슈 => Axios withCredential true     AuthForm 스타일 변경     MyPageBox 스타일 변경 todo:     MyPageOption(좋아요, 내가 쓴 글) 에 따른 렌더링 수정 필요     BoardSimple Column 축소 필요

+++ sunbro/src/main/java/com/humorpage/sunbro/service/EmailAuthService.java
@@ -52,16 +52,20 @@ public class EmailAuthService {
     //redis에 expire time 정해서 보냄
     @Async("emailAuthExecutor")
     public void sendMailwithKey(String email) throws Exception{
-        String key = RandomGenerator.RandomnameGenerate(8);
-        try{
-            redisTokenService.setDataExpire(email,key,JwtTokenService.EmailAuthValidSecond);
-            Date expires = new Date(System.currentTimeMillis()+JwtTokenService.EmailAuthValidMilisecond);
-            MimeMessage message = createAuthMessage(email,key,emailTimeFormat.format(expires));
-            emailSender.send(message);
-        }catch (Exception e){
-            e.printStackTrace();
-            redisTokenService.deleteData(email);
-            throw e;
+        String key;
+        key = redisTokenService.getData(email);
+        if(key==null){
+            key = RandomGenerator.RandomnameGenerate(8);
+            try{
+                redisTokenService.setDataExpire(email,key,JwtTokenService.EmailAuthValidSecond);
+                Date expires = new Date(System.currentTimeMillis()+JwtTokenService.EmailAuthValidMilisecond);
+                MimeMessage message = createAuthMessage(email,key,emailTimeFormat.format(expires));
+                emailSender.send(message);
+            }catch (Exception e){
+                e.printStackTrace();
+                redisTokenService.deleteData(email);
+                throw e;
+            }
         }
     }
 }
